+++ docs/README.md
@@ -39,11 +39,12 @@
 - [x] command dto 방식 적용
   - [ ] start 안하고 바로 move 오류 수정
   - [ ] 인덴트 줄이기
+- [ ] 전체 흐름 따라가보기
 - [ ] 리뷰 반영
-- [ ] view에 넘겨줄 때 dto 방식 고려
 - [ ] direction vector 방식 적용
 - [ ] move 메서드 개선
 - [ ] 우리 프로그램의 장점 고려, 유지
+- [ ] view에 넘겨줄 때 dto 방식 고려
 
 ## 체스 말 이동 규칙
