+++ src/main/java/chess/dto/CommandDto.java
@@ -1,4 +1,4 @@
-package chess.controller;
+package chess.dto;
 
 import chess.domain.Command;
 import java.util.Arrays;
