+++ src/test/java/chess/domain/CommandTest.java
@@ -2,7 +2,7 @@ package chess.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.controller.CommandDto;
+import chess.dto.CommandDto;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
